+++ Data/Models/Product.cs
@@ -6,11 +6,11 @@ namespace ThumbsUpGroceries_backend.Data.Models
     {
         public int ProductId { get; set; }
         public string Name { get; set; }
-        public float Price { get; set; }
+        public int Price { get; set; }
         public PriceUnitType PriceUnitType { get; set; }
         public string? Description { get; set; }
         public string? Images { get; set; }
-        public float Quantity { get; set; }
+        public int Quantity { get; set; }
         public List<int> Categories { get; set; }
         public float? Rating { get; set; }
         public int? ReviewCount { get; set; }
@@ -20,11 +20,11 @@ namespace ThumbsUpGroceries_backend.Data.Models
     {
         public int ProductId { get; set; }
         public string Name { get; set; } = string.Empty;
-        public float Price { get; set; }
+        public int Price { get; set; }
         public PriceUnitType PriceUnitType { get; set; } = PriceUnitType.ea;
         public string? Description { get; set; }
         public List<string>? Images { get; set; }
-        public float Quantity { get; set; }
+        public int Quantity { get; set; }
         public List<int> Categories { get; set; } = [];
         public float? Rating { get; set; }
         public int? ReviewCount { get; set; }
@@ -34,7 +34,7 @@ namespace ThumbsUpGroceries_backend.Data.Models
     {
         public int ProductId { get; set; }
         public string Name { get; set; } = string.Empty;
-        public float Price { get; set; }
+        public int Price { get; set; }
         public PriceUnitType PriceUnitType { get; set; } = PriceUnitType.ea;
         public string Image { get; set; } = string.Empty;
         public float Rating { get; set; }
@@ -47,13 +47,13 @@ namespace ThumbsUpGroceries_backend.Data.Models
         [StringLength(255)]
         public string Name { get; set; }
         [Required]
-        public float? Price { get; set; }
+        public int? Price { get; set; }
         [Required]
         public PriceUnitType PriceUnitType { get; set; }
         public string? Description { get; set; }
         public List<IFormFile>? Images { get; set; }
         [Required]
-        public float Quantity { get; set; }
+        public int Quantity { get; set; }
         [Required]
         public List<int> Categories { get; set; }
     }
@@ -62,12 +62,12 @@ namespace ThumbsUpGroceries_backend.Data.Models
     {
         [StringLength(255)]
         public string? Name { get; set; }
-        public float? Price { get; set; }
+        public int? Price { get; set; }
         public PriceUnitType? PriceUnitType { get; set; }
         public string? Description { get; set; }
         public List<IFormFile>? Images { get; set; }
-        public float? Quantity { get; set; }
-        public float? AddedQuantity { get; set; }
+        public int? Quantity { get; set; }
+        public int? AddedQuantity { get; set; }
         public List<int>? Categories { get; set; }
     }
 }
